feat: implement theme support for the Wold Emoji Day noodle
