+++ app/components/Noodle/EmojiDay/Logo.vue
@@ -1,18 +1,12 @@
 <template>
-  <div ref="sizer" class="sizer w-full sm:w-[540px] md:w-[640px] lg:w-[900px]"></div>
+  <div class="relative flex">
+    <ThemedLogo v-if="colorMode.value === 'dark'" :emojiSets="darkTheme" />
+    <ThemedLogo v-else :emojiSets="lightTheme" />
+  </div>
 </template>
 
-<style>
-.sizer {
-  position: relative;
-  height: auto;
-  aspect-ratio: 900/360;
-}
-</style>
-
 <script setup lang="ts">
-import { init } from './emoji-thing'
-import { onMounted, onUnmounted, useTemplateRef } from 'vue'
+import ThemedLogo from './ThemedLogo.vue'
 import red from './emojisets/red.png'
 import blue from './emojisets/blue.png'
 import faces from './emojisets/faces.png'
@@ -20,9 +14,12 @@ import gold from './emojisets/gold.png'
 import hearts from './emojisets/hearts.png'
 import green from './emojisets/green.png'
 import earth from './emojisets/earth.png'
-import cloud from './emojisets/cloud.png'
+import light from './emojisets/themed-light.png'
+import dark from './emojisets/themed-dark.png'
 
-const EMOJI_SET_SOURCES = {
+const colorMode = useColorMode()
+
+const lightTheme = {
   red,
   blue,
   hearts,
@@ -30,86 +27,11 @@ const EMOJI_SET_SOURCES = {
   faces,
   green,
   earth,
-  cloud,
+  themed: light,
 }
 
-let emojiSetImagesPromise: undefined | Promise<Record<string, HTMLImageElement>>
-
-function loadEmojiSetImages() {
-  if (emojiSetImagesPromise) {
-    return emojiSetImagesPromise
-  }
-
-  const promise = Promise.all(
-    Object.entries(EMOJI_SET_SOURCES).map(([name, src]) => {
-      return new Promise<[string, HTMLImageElement]>((resolve, reject) => {
-        const image = new Image()
-        image.addEventListener('load', () => {
-          resolve([name, image])
-        })
-        image.addEventListener('error', () => {
-          if (promise === emojiSetImagesPromise) {
-            emojiSetImagesPromise = undefined
-          }
-          reject(new Error(`could not load image ${JSON.stringify(src)}`))
-        })
-        image.src = src
-      })
-    }),
-  ).then(sets => Object.fromEntries(sets))
-
-  emojiSetImagesPromise = promise
-  return emojiSetImagesPromise
+const darkTheme = {
+  ...lightTheme,
+  themed: dark,
 }
-
-let running = false
-let handle: undefined | ReturnType<typeof init>
-let observer: undefined | IntersectionObserver
-
-const sizerRef = useTemplateRef<HTMLDivElement>('sizer')
-
-onMounted(() => {
-  const sizer = sizerRef.value
-  if (!sizer) {
-    return
-  }
-
-  observer = new IntersectionObserver(
-    ([entry]) => {
-      if (!entry) {
-        return
-      }
-
-      if (entry.isIntersecting) {
-        running = true
-        loadEmojiSetImages().then(sets => {
-          if (!running) {
-            return
-          }
-          if (!handle) {
-            handle = init(sizer, sets)
-          }
-          handle.start()
-        })
-      } else {
-        running = false
-        handle?.pause()
-      }
-    },
-    {
-      rootMargin: '100px',
-    },
-  )
-  observer.observe(sizer)
-})
-
-onUnmounted(() => {
-  handle?.destroy()
-  handle = undefined
-
-  observer?.disconnect()
-  observer = undefined
-
-  running = false
-})
 </script>
