+++ app/components/Noodle/EmojiDay/emojisets/themed-dark.png
